add modules and webpack
