+++ src/constants.js
@@ -1,0 +1,9 @@
+export const FILTER_NAMES = [
+  `All`,
+  `Overdue`,
+  `Today`,
+  `Favorites`,
+  `Repeating`,
+  `Tags`,
+  `Archive`,
+];
