+++ src/utils.js
@@ -1,0 +1,2 @@
+export const getRandomNumber = (min, max) =>
+  Math.round(min - 0.5 + Math.random() * (max - min + 1));
